+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [Unreleased]
 ### Fixed
-- Resource Finder - Set a Sequelize bug workaround to get the WHERE clause supported when no includes are available.
+- Resource Retrieval - Fix the record retrieval for a record having no belongsTo/hasOne for projects using Sequelize 4.8.x.
 
 ## RELEASE 1.3.6 - 2017-09-10
 ### Changed