+++ CHANGELOG.md
@@ -3,10 +3,12 @@
 ## [Unreleased]
 ### Added
 - Line Charts - Support SQLite dialect for groupBy fields.
+- Smart Fields - Add a parameter to specify if the sorting is allowed on this field.
 
 ### Fixed
 - Resource Retrieval - Fix the record retrieval for a record having no belongsTo/hasOne for projects using Sequelize 4.8.x.
 - Charts - Fix broken charts if they contain relationship filters on model fields that have a different column name.
+- Initialization - Ignore directories while loading models.
 
 ## RELEASE 1.3.6 - 2017-09-10
 ### Changed